requirements.txt
requests~=2.28.1
Flask~=2.2.2
mysql-connector-python~=8.0.31
Flask-Cors~=3.0.10
py-zabbix~=1.1.7